src/bot__db/requirements.txt
flask>=2.3.2
flask-admin>=1.6.1
flask-sqlalchemy>=3.0.5
Flask-Migrate>=4.0.4
SQLAlchemy-serializer>=1.4.1
pydantic>=1.10.9